Develop

+++ src/main/java/fr/eni/encheres/dal/CategorieDAO.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 public interface CategorieDAO {
     public void create(Categorie categorie);
-    public void delete(int idCategorie);
+    public void delete(long idCategorie);
     public List<Categorie> readAll();
 }
