+++ src/main/java/fr/eni/encheres/dal/CategorieDAOImpl.java
@@ -5,13 +5,14 @@ import fr.eni.encheres.bo.Categorie;
 import java.util.List;
 
 public class CategorieDAOImpl implements CategorieDAO {
+
     @Override
     public void create(Categorie categorie) {
 
     }
 
     @Override
-    public void delete(int idCategorie) {
+    public void delete(long idCategorie) {
 
     }
