+++ src/main/java/fr/eni/encheres/dal/EnchereDAO.java
@@ -6,10 +6,10 @@ import fr.eni.encheres.bo.Enchere;
 import java.util.List;
 
 public interface EnchereDAO {
-    public void create(int idArticle);
+    public void create(long idArticle);
     public List<Enchere> findAll();
     public Enchere findByEtat(String etat);
-    public void delete(int idEnchere);
+    public void delete(long idEnchere);
     public Enchere findByNom(String nom);
-    public List<Enchere> findByCategorie(int idCategorie);
+    public List<Enchere> findByCategorie(long idCategorie);
 }
