+++ src/main/java/fr/eni/encheres/dal/RetraitDAO.java
@@ -3,6 +3,6 @@ package fr.eni.encheres.dal;
 import fr.eni.encheres.bo.Retrait;
 
 public interface RetraitDAO {
-    public Retrait read(int idArticle);
+    public Retrait read(long idArticle);
     public void create(Retrait retrait);
 }
